requirements.txt
asgiref==3.7.2
Django==4.2.1
django-taggit==5.0.1
flake8==6.1.0
mccabe==0.7.0
Pillow==10.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
